Cost distance

Calculate the least-cost distance between points.

costDistance(x, fromCoords, toCoords)

Arguments

x
object of class TransitionLayer

fromCoords
first set of point locations (of SpatialPoints, matrix or numeric class)

toCoords
optional, second set of point locations (of SpatialPoints, matrix or numeric class)

Value

distance matrix (S3 class dist or matrix)

Details

Cost units between cells are defined as the reciprocal of the values in the transition matrix.

The function uses Dijkstra's algorithm, as implemented in the igraph package.

A projection correction is needed for accuracy in the case of grid data for a longlat raster (see function geoCorrection).

References

E.W. Dijkstra. 1959. A note on two problems in connexion with graphs. Numerische Mathematik 1, 269 - 271.

See also

Examples

library("raster")
# create a new raster and set all its values to unity.
r <- raster(nrows=18, ncols=36)
r <- setValues(r,runif(ncell(r),0,1))

# create a Transition object from the raster
tr <- transition(r,function(x) 1/mean(x),8)

# asymmetric
ncf <- function(x) max(x) - x[1] + x[2]
tr2 <- transition(r,ncf,8, symm=FALSE)

# create two sets of coordinates
sP1 <- cbind(c(65,5,-65),c(55,35,-35))
sP2 <- cbind(c(50,15,-40),c(80,20,-5))

# from and to identical
costDistance(tr,sP1)

#>          1        2
#> 2 1.398406         
#> 3 2.927121 2.362693

costDistance(tr2,sP1)

#>           [,1]     [,2]     [,3]
#> [1,]  0.000000 6.444509 89.76445
#> [2,]  8.071019 0.000000 83.39793
#> [3,] 16.262011 8.465638  0.00000

# from and to different
costDistance(tr,sP1,sP2)

#>           [,1]     [,2]     [,3]
#> [1,] 0.6707114 1.467142 2.811153
#> [2,] 1.5863626 1.048030 2.357321
#> [3,] 3.1150776 2.147374 1.078289

costDistance(tr2,sP1,sP2)

#>           [,1]    [,2]      [,3]
#> [1,]  3.207926 5.68906 11.634261
#> [2,]  6.223955 2.10860  5.267742
#> [3,] 14.689593 9.11994  3.379581